Updates and P1 changes

+++ .systemfiles/install.sh
@@ -4,16 +4,16 @@
 
 install() {
     if [[ -f "$1" ]] && sudo diff "$1" "$HOME/.systemfiles$1" >/dev/null; then
-        echo "$1: Ok"
+        echo "[OK] $1"
     else
         if [[ -f "$1" ]]; then
             sudo diff -u "$1" "$HOME/.systemfiles$1"
-            if confirm "$1: replace?"; then
+            if confirm "[Change] $1: replace?"; then
                 sudo mkdir -p "$(dirname "$1")"
                 sudo cp "$HOME/.systemfiles$1" "$1"
             fi
         else
-            echo "$1 missing, copying..."
+            echo "[Change] $1 missing, copying..."
             sudo mkdir -p "$(dirname "$1")"
             sudo cp "$HOME/.systemfiles$1" "$1"
         fi
@@ -21,7 +21,10 @@ install() {
 }
 
 remove() {
-    sudo rm -f "$1"
+    if [[ -f "$1" ]]; then
+        echo "[Remove] $1"
+        sudo rm -f "$1"
+    fi
 }
 
 install "/etc/default/keyboard"
@@ -32,7 +35,8 @@ install "/etc/udev/rules.d/46-TI_launchpad.rules"
 install "/etc/udev/rules.d/47-altera.rules"
 install "/etc/udev/rules.d/60-vboxdrv.rules"
 install "/etc/udev/rules.d/85-tessel.rules"
-install "/etc/udev/rules.d/backlight.rules"
+remove "/etc/udev/rules.d/backlight.rules"
+install "/etc/udev/rules.d/90-backlight.rules"
 install "/etc/modprobe.d/thinkpad_acpi.conf"
 install "/usr/share/xsessions/custom.desktop"
 
@@ -50,6 +54,7 @@ if [[ $(hostname -s) =~ juho-laptop ]]; then
     # P1 G3
     if grep -q i9-10885H /proc/cpuinfo; then
         install "/etc/thinkfan.yaml"
+        install "/etc/tlp.d/100-nvme.conf"
     fi
 
     install "/etc/NetworkManager/dispatcher.d/99nfs"
